TINY-10237: Moved the Sanitization module and added basic test
